Modify JSON parser to enable arbitrary object creation during parsing
This change makes it possible to greatly reduce memory usage when parsing large arrays of objects. As each array member is parsed into an object, any unused data from the JSON source can be discared immediately after the object is generated (and before parsing the next object). For instance, when parsing an array of tweets, we can now generate the Tweet objects directly and discard any extra data from the JSON source immediately before moving on to the next object in the JSON stream.